backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
boto3==1.35.45
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
orjson==3.10.7
loguru==0.7.2
torch==2.4.1
trimesh==4.4.9
numpy==2.0.1
pygltflib==1.16.0
